+++ internal/acceptance_tests/blocks/logging_scalyr_nested_placement_none.tf
@@ -1,0 +1,7 @@
+logging_scalyr {
+  name = "{{.LOGGING_SCALYR_NAME}}"
+  authentication = {
+    token = "test-scalyr-token"
+  }
+  placement = "none"
+}